// src/common/common.enum.ts
export enum LanguageEnum {
  VI = 'vi',
  EN = 'en'
}

export enum RouteEnum {
  WILDCARD_PATH = '/*',

  LOGIN = '/login',
  HOME = '/',
  CATALOGUE = '/catalogue',
  COURSE = '/course',
  COURSE_DETAILS = '/course-details',
  BLOG = '/blog',
  EVENT = '/event',
  INFO = '/info',
  PROFILE = '/profile',
  SEARCH = '/search',

  ADMIN = '/admin'
}